+++ src/main/java/net/Abdymazhit/VimeReplays/replay/data/UnsneakingData.java
@@ -6,11 +6,11 @@ public class UnsneakingData extends RecordingData implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private final int entityId;
+    private final short entityId;
 
-    public UnsneakingData(int entityId) {
+    public UnsneakingData(short entityId) {
         this.entityId = entityId;
     }
 
-    public int getEntityId() { return entityId; }
+    public short getEntityId() { return entityId; }
 }
